+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## [Unreleased](https://github.com/Amebis/eduVPN/compare/3.3...HEAD)
 
 - Increase minimum window height
+- Show update button on the About page when available
 
 
 ## [3.3](https://github.com/Amebis/eduVPN/compare/3.2.2...3.3) (2022-08-26)